Refactor A3DControler and A3DModel classes
This commit refactors the A3DControler and A3DModel classes in the a3dcontroler.py and a3dmodel.py files respectively. The changes include updating the temperature and max_tokens parameters in the A3DControler class, and adding fine_temperature and fine_max_tokens parameters in the A3DModel class.